// types/enums.ts
export enum ContentDataType {
  VIDEO = "video",
  AUDIO = "audio",
  IMAGE = "image",
}

export enum ContentSourceType {
  EDEDUN = "ededun",
  NEW = "new",
}

export enum Level {
  BUILDER = "builder",
  FOUNDATION = "foundation",
  EXPLORER = "explorer",
}

export enum LanguageCode {
  // ENGLISH = "EN",
  // SPANISH = "ES",
  // FRENCH = "FR",
  // GERMAN = "DE",
  // ITALIAN = "IT",
  // PORTUGUESE = "PT",
  // MANDARIN = "ZH",
  // JAPANESE = "JA",
  // KOREAN = "KO",
  // ARABIC = "AR",
  // RUSSIAN = "RU",
  // HINDI = "HI",
  YORUBA = "YO",
  // IGBO = "IG",
  // HAUSA = "HA",
  // SWAHILI = "SW",
}